Rename directory from `deployments` to `projects`
