+++ .gitignore
@@ -49,7 +49,6 @@ profile.log
 saved-state.json
 last-query-output.json
 staging-databases
-projects/
 /rilldata
 docs/dist/
