+++ web-admin/src/components/projects/ProjectBuilding.svelte
